Started work on updating the Huffman Tree when a character is inserted.
The Huffman Tree will now update the character node for the inserted
character. If there is no node for the character, a new tree node will
be created. The new character (along with the new empty node) will be
inserted into the linked list that's connecting the tree nodes. If a
node is already present for the character, its Frequency will be
incremented.

At this point, the tree still needs to be re-balanced.