Fixed the problem with adding characters to the Huffman Tree class
After adding in the use of the ID check and going down the tree
manually, I found a data loss in the tree portion of the structure. I
fixed it, and ran the test in the Program class. I found that the
resulting linked list contains nodes that formed a tree that fall
with-in all the boundaries of what I'm supposed to get.