Going to the next step in the project.
Now that I've got an Adaptive Huffman Tree class for encoding and
decoding char characters, I'm going to build proper encoders and
decoders that rely on that class. The tree is convert a char to a bit
array and back, but we need something that works with bytes, that can
write/read to/from files, and stuff like that.