Fixed the reader error that was causing an infinate loop in the reader.
When it came time to read the bits from the BitArray into the bool[], I
only counted up to the number of bytes. I should have counted up to the
number of bits. I forgot to multiply the bytesRead by the number of bits
per byte.

I also cut down the number of bytes that get inserted into the BitArray.