I'm throwing in my old C++ code from the school project that introduced me to Adaptive Huffman Encoding.
I created a project for throwing in my old files, so there a a lot of
extra files were added. The school project was to be compiled on a linux
computer by the teacher. He gave us the initial definition for the
AdaptiveHuffman.h file (on paper), and we were to implement it.

These are the files I turned in:
AdaptiveHuffman.h
AdaptiveHuffman.cpp
CharList.h
CharList.cpp